Modified SwingProgressFunction due to new JGraLab interface (int -> long)
git-svn-id: https://svn.uni-koblenz.de/gupro/re-group/trunk/project/jgstreetmap@6236 6c03965d-96ef-0310-af27-ce2fa9dd3b5a

+++ src/de/uni_koblenz/jgstreetmap/gui/SwingProgressFunction.java
@@ -20,7 +20,7 @@ public class SwingProgressFunction implements ProgressFunction, ActionListener {
 	private JProgressBar pb;
 	private String title;
 	private String label;
-	private int steps;
+	private long steps;
 	private BoundedRangeModel brm;
 	private JLabel lbl;
 	private long startTime;
@@ -40,12 +40,12 @@ public class SwingProgressFunction implements ProgressFunction, ActionListener {
 	}
 
 	@Override
-	public int getInterval() {
+	public long getInterval() {
 		return brm.getMaximum() > steps ? 1 : steps / brm.getMaximum();
 	}
 
 	@Override
-	public void init(int steps) {
+	public void init(long steps) {
 		this.steps = steps;
 		wnd = new JFrame(title);
 		wnd.setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
@@ -78,7 +78,7 @@ public class SwingProgressFunction implements ProgressFunction, ActionListener {
 	}
 
 	@Override
-	public void progress(int progress) {
+	public void progress(long progress) {
 		if (brm.getValue() < brm.getMaximum()) {
 			brm.setValue(brm.getValue() + 1);
 			setTimeText();
